Prevent loading scripts multiple times per process
Previously, scripts would be loaded whenever Sidekiq::Grouping::Redis
was initialized. The assumption was that this would only be initialized
once per process, but that turned out to be an incorrect assumption. We
need to make sure that even if the class is instantiated many times,
scripts are only loaded once. We also want to make sure that we still
lazy-load scripts so that we don't do this during server initialization
if it's not needed.

To implement these requirements, we simply use a Concurrent::Map from
concurrent-ruby. It has a compute_if_absent method available that we can
use to lazily load all scripts in a thread-safe way.