Prevent loading scripts multiple times per process
A port of this commit which changes were originaly omitted from
a recent merge conflict resolution.
https://github.com/ParentSquare/sidekiq-grouping/commit/d82e1a0a8a0b6bcc8ee02605511ae23af1686677

The existing code loads all redis scripts during application load
however that means we're waiting for requests to finish during
application load which creates latency and potentially errors in
boostrapping. Rather, we use the same mechanism from this other
commit to lazy load redis scripts once and only once in a thread
safe manner, so they are loaded once needed and not again.